+++ Demo/DemoTests/MoyaProviderIntegrationTests.swift
@@ -263,28 +263,6 @@ class MoyaProviderIntegrationTests: QuickSpec {
 
                         expect{message}.toEventually( equal(userMessage) )
                     }
-                    
-                    it("returns identical signals for inflight requests") {
-                        let target: GitHub = .Zen
-                        
-                        expect(provider.inflightRequests.count).to(equal(0))
-                        
-                        var receivedResponse: MoyaResponse!
-                        
-                        provider.request(target).subscribeNext { (response) -> Void in
-                            receivedResponse = response as? MoyaResponse
-                            expect(provider.inflightRequests.count).to(equal(1))
-                        }
-                        
-                        provider.request(target).subscribeNext { (response) -> Void in
-                            expect(receivedResponse).toNot(beNil())
-                            expect(receivedResponse).to(beIndenticalToResponse(response as! MoyaResponse))
-                            expect(provider.inflightRequests.count).to(equal(1))
-                        }
-                        
-                        // Allow for network request to complete
-                        expect(provider.inflightRequests.count).toEventually( equal(0) )
-                    }
                 }
             }
             
@@ -314,28 +292,6 @@ class MoyaProviderIntegrationTests: QuickSpec {
                     }
                     
                     expect{message}.toEventually( equal(userMessage) )
-                }
-                
-                it("returns identical signals for inflight requests") {
-                    let target: GitHub = .Zen
-                    
-                    expect(provider.inflightRequests.count).to(equal(0))
-                    
-                    var receivedResponse: MoyaResponse!
-                    
-                    provider.request(target).startWithNext { (response) -> Void in
-                        receivedResponse = response
-                        expect(provider.inflightRequests.count).to(equal(1))
-                    }
-                    
-                    provider.request(target).startWithNext { (response) -> Void in
-                        expect(receivedResponse).toNot(beNil())
-                        expect(receivedResponse).to(beIndenticalToResponse(response))
-                        expect(provider.inflightRequests.count).to(equal(1))
-                    }
-                    
-                    // Allow for network request to complete
-                    expect(provider.inflightRequests.count).toEventually( equal(0) )
                 }
             }
         }