+++ Demo/DemoTests/MoyaProviderIntegrationTests.swift
@@ -41,6 +41,35 @@ extension GitHub : MoyaTarget {
     }
 }
 
+private enum HTTPBin: MoyaTarget {
+    case BasicAuth
+
+    var baseURL: NSURL { return NSURL(string: "http://httpbin.org")! }
+    var path: String {
+        switch self {
+        case .BasicAuth:
+            return "/basic-auth/user/passwd"
+        }
+    }
+
+    var method: Moya.Method {
+        return .GET
+    }
+    var parameters: [String: AnyObject]? {
+        switch self {
+        default:
+            return [:]
+        }
+    }
+
+    var sampleData: NSData {
+        switch self {
+        case .BasicAuth:
+            return "{\"authenticated\": true, \"user\": \"user\"}".dataUsingEncoding(NSUTF8StringEncoding)!
+        }
+    }
+}
+
 private func url(route: MoyaTarget) -> String {
     return route.baseURL.URLByAppendingPathComponent(route.path).absoluteString
 }
@@ -79,6 +108,11 @@ class MoyaProviderIntegrationTests: QuickSpec {
             OHHTTPStubs.stubRequestsPassingTest({$0.URL!.path == "/users/ashfurrow"}) { _ in
                 return OHHTTPStubsResponse(data: GitHub.UserProfile("ashfurrow").sampleData, statusCode: 200, headers: nil).responseTime(0.5)
             }
+            
+            OHHTTPStubs.stubRequestsPassingTest({$0.URL!.path == "/basic-auth/user/passwd"}) { _ in
+                return OHHTTPStubsResponse(data: HTTPBin.BasicAuth.sampleData, statusCode: 200, headers: nil)
+            }
+            
         }
 
         afterEach {
@@ -130,6 +164,41 @@ class MoyaProviderIntegrationTests: QuickSpec {
                         token.cancel()
 
                         expect(receivedError).toEventuallyNot( beNil() )
+                    }
+                }
+                
+                describe("a provider with credential closures") {
+                    it("credential closure returns nil") {
+                        var called = false
+                        let provider  = MoyaProvider<HTTPBin>(credentialClosure: {(target) -> (NSURLCredential?) in
+                            
+                            called = true
+                            return nil
+                        })
+                        
+                        let target: HTTPBin = .BasicAuth
+                        provider.request(target) { (data, statusCode, response, error) in }
+                        
+                        expect(called).toEventually( beTrue() )
+                        
+                    }
+                    
+                    it("credential closure returns valid username and password") {
+                        var called = false
+                        var returnedData: NSData?
+                        let provider  = MoyaProvider<HTTPBin>(credentialClosure: {(target) -> (NSURLCredential?) in
+                            
+                            called = true
+                            return NSURLCredential(user: "user", password: "passwd", persistence: .None)
+                        })
+                        
+                        let target: HTTPBin = .BasicAuth
+                        provider.request(target) { (data, statusCode, response, error) in
+                            returnedData = data
+                        }
+                        
+                        expect(called).toEventually( beTrue() )
+                        expect(returnedData).toEventually(equal(target.sampleData))
                     }
                 }
 
@@ -193,30 +262,6 @@ class MoyaProviderIntegrationTests: QuickSpec {
                         }
 
                         expect{message}.toEventually( equal(userMessage) )
-                    }
-                    
-                    it("returns identical signals for inflight requests") {
-                        let target: GitHub = .Zen
-                        let signal1 = provider.request(target)
-                        let signal2 = provider.request(target)
-                        
-                        expect(provider.inflightRequests.count).to(equal(0))
-                        
-                        var receivedResponse: MoyaResponse!
-                        
-                        signal1.subscribeNext { (response) -> Void in
-                            receivedResponse = response as? MoyaResponse
-                            expect(provider.inflightRequests.count).to(equal(1))
-                        }
-                        
-                        signal2.subscribeNext { (response) -> Void in
-                            expect(receivedResponse).toNot(beNil())
-                            expect(receivedResponse).to(beIndenticalToResponse( response as! MoyaResponse) )
-                            expect(provider.inflightRequests.count).to(equal(1))
-                        }
-                        
-                        // Allow for network request to complete
-                        expect(provider.inflightRequests.count).toEventually( equal(0) )
                     }
                 }
             }